backend/requirements.txt
livekit-agents
livekit
livekit-plugins-silero
livekit-plugins-google
langgraph
python-dotenv
httpx
psycopg[binary]
